+++ client/src/components/settings/index.js
@@ -2,6 +2,7 @@ export { default as DatabaseManagement } from './DatabaseManagement';
 export { default as BOMSettings } from './BOMSettings';
 export { default as SMTPSettings } from './SMTPSettings';
 export { default as ECOSection } from './ECOSection';
+export { default as ECOSettingsTab } from './ECOSettingsTab';
 export { default as UserManagement } from './UserManagement';
 export { default as CategorySpecificationsManager } from './CategorySpecificationsManager';
 export { default as ApprovalStagesSettings } from './ApprovalStagesSettings';